Remove obsolete KML archive implementation
