+++ pkg/api/handlers.go
@@ -14,7 +14,8 @@ import (
 	"time"
 
 	"chicha-isotope-map/pkg/database"
-	"chicha-isotope-map/pkg/kmlarchive"
+	"chicha-isotope-map/pkg/jsonarchive"
+	"chicha-isotope-map/pkg/trackjson"
 )
 
 // =======================
@@ -27,18 +28,13 @@ import (
 type Handler struct {
 	DB      *database.Database
 	DBType  string
-	Archive *kmlarchive.Generator
+	Archive *jsonarchive.Generator
 	Logf    func(string, ...any)
 }
 
-// microRoentgenPerMicroSievert stores the factor for translating between
-// microsieverts per hour and microroentgen per hour. We keep it in one place so
-// both the exporter and importer remain consistent with each other.
-const microRoentgenPerMicroSievert = 100.0
-
 // NewHandler constructs a Handler with sane defaults.
 // Logf is optional; pass nil if logging is not required.
-func NewHandler(db *database.Database, dbType string, archive *kmlarchive.Generator, logf func(string, ...any)) *Handler {
+func NewHandler(db *database.Database, dbType string, archive *jsonarchive.Generator, logf func(string, ...any)) *Handler {
 	return &Handler{DB: db, DBType: dbType, Archive: archive, Logf: logf}
 }
 
@@ -53,7 +49,7 @@ func (h *Handler) Register(mux *http.ServeMux) {
 	mux.HandleFunc("/api/tracks/months/", h.handleTracksByMonth)
 	mux.HandleFunc("/api/track/", h.handleTrackData)
 	mux.HandleFunc("/api/tracks/", h.handleTrackData) // legacy alias for older clients
-	mux.HandleFunc("/api/kml/daily.tar.gz", h.handleArchiveDownload)
+	mux.HandleFunc("/api/json/daily.tar.gz", h.handleArchiveDownload)
 }
 
 // handleOverview publishes machine-readable docs so developers understand
@@ -74,7 +70,7 @@ func (h *Handler) handleOverview(w http.ResponseWriter, r *http.Request) {
 		LatestTrackIndex int64             `json:"latestTrackIndex"`
 		LatestTrackID    string            `json:"latestTrackID,omitempty"`
 	}{
-		Disclaimers:      disclaimerTexts,
+		Disclaimers:      trackjson.Disclaimers,
 		TotalTracks:      totalTracks,
 		LatestTrackIndex: totalTracks,
 		LatestTrackID:    latestTrackID,
@@ -108,10 +104,10 @@ func (h *Handler) handleOverview(w http.ResponseWriter, r *http.Request) {
 				"query":       []string{"startAfter", "limit"},
 				"description": "Lists tracks for a calendar month using the same pagination fields.",
 			},
-			"dailyKML": map[string]any{
+			"dailyJSON": map[string]any{
 				"method":      "GET",
-				"path":        "/api/kml/daily.tar.gz",
-				"description": "Downloads the current tar.gz bundle of all published KML files.",
+				"path":        "/api/json/daily.tar.gz",
+				"description": "Downloads the current tar.gz bundle of all published .cim JSON tracks.",
 				"frequency":   "Updated once per day",
 			},
 		},
@@ -179,7 +175,7 @@ func (h *Handler) handleTracksList(w http.ResponseWriter, r *http.Request) {
 		NextStartAfter: next,
 		TotalTracks:    totalTracks,
 		LatestTrackID:  latestTrackID,
-		Disclaimers:    disclaimerTexts,
+		Disclaimers:    trackjson.Disclaimers,
 	}
 
 	h.respondJSON(w, resp)
@@ -334,7 +330,7 @@ func (h *Handler) handleTracksByYear(w http.ResponseWriter, r *http.Request) {
 		RangeTotal:     rangeTotal,
 		TotalTracks:    totalTracks,
 		LatestTrackID:  latestTrackID,
-		Disclaimers:    disclaimerTexts,
+		Disclaimers:    trackjson.Disclaimers,
 	}
 
 	h.respondJSON(w, resp)
@@ -441,31 +437,10 @@ func (h *Handler) handleTracksByMonth(w http.ResponseWriter, r *http.Request) {
 		RangeTotal:     rangeTotal,
 		TotalTracks:    totalTracks,
 		LatestTrackID:  latestTrackID,
-		Disclaimers:    disclaimerTexts,
+		Disclaimers:    trackjson.Disclaimers,
 	}
 
 	h.respondJSON(w, resp)
-}
-
-// serveTrackData centralises the marker streaming logic so both ID-based and
-// index-based handlers produce identical responses.
-type trackMarkerPayload struct {
-	ID                     int64    `json:"id"`
-	TimeUnix               int64    `json:"timeUnix"`
-	TimeUTC                string   `json:"timeUTC"`
-	Lat                    float64  `json:"lat"`
-	Lon                    float64  `json:"lon"`
-	AltitudeM              *float64 `json:"altitudeM,omitempty"`
-	DoseRateMicroSvH       float64  `json:"doseRateMicroSvH"`
-	DoseRateMicroRoentgenH float64  `json:"doseRateMicroRh"`
-	CountRateCPS           float64  `json:"countRateCPS"`
-	SpeedMS                float64  `json:"speedMS"`
-	SpeedKMH               float64  `json:"speedKMH"`
-	TemperatureC           *float64 `json:"temperatureC,omitempty"`
-	HumidityPercent        *float64 `json:"humidityPercent,omitempty"`
-	DetectorName           string   `json:"detectorName,omitempty"`
-	DetectorType           string   `json:"detectorType,omitempty"`
-	RadiationTypes         []string `json:"radiationTypes,omitempty"`
 }
 
 func (h *Handler) serveTrackData(w http.ResponseWriter, r *http.Request, trackID string) {
@@ -514,7 +489,7 @@ func (h *Handler) serveTrackData(w http.ResponseWriter, r *http.Request, trackID
 			capEstimate = int(summary.MarkerCount)
 		}
 	}
-	markers := make([]trackMarkerPayload, 0, capEstimate)
+	markers := make([]trackjson.MarkerPayload, 0, capEstimate)
 
 	nextID := from
 	for nextID <= to {
@@ -538,7 +513,8 @@ func (h *Handler) serveTrackData(w http.ResponseWriter, r *http.Request, trackID
 		for marker := range markersCh {
 			got = true
 			lastID = marker.ID
-			markers = append(markers, makeTrackMarkerPayload(marker))
+			payload, _ := trackjson.MakeMarkerPayload(marker)
+			markers = append(markers, payload)
 		}
 
 		if err := <-errCh; err != nil {
@@ -554,71 +530,32 @@ func (h *Handler) serveTrackData(w http.ResponseWriter, r *http.Request, trackID
 		nextID = lastID + 1
 	}
 
-	w.Header().Set("Content-Disposition", fmt.Sprintf("attachment; filename=%s", safeCIMFilename(trackID)))
+	w.Header().Set("Content-Disposition", fmt.Sprintf("attachment; filename=%s", trackjson.SafeCIMFilename(trackID)))
 
 	resp := struct {
-		TrackID     string               `json:"trackID"`
-		TrackIndex  int64                `json:"trackIndex"`
-		APIURL      string               `json:"apiURL"`
-		FirstID     int64                `json:"firstID"`
-		LastID      int64                `json:"lastID"`
-		MarkerCount int64                `json:"markerCount"`
-		Markers     []trackMarkerPayload `json:"markers"`
-		Disclaimers map[string]string    `json:"disclaimers"`
+		TrackID     string                    `json:"trackID"`
+		TrackIndex  int64                     `json:"trackIndex"`
+		APIURL      string                    `json:"apiURL"`
+		FirstID     int64                     `json:"firstID"`
+		LastID      int64                     `json:"lastID"`
+		MarkerCount int64                     `json:"markerCount"`
+		Markers     []trackjson.MarkerPayload `json:"markers"`
+		Disclaimers map[string]string         `json:"disclaimers"`
 	}{
 		TrackID:     trackID,
 		TrackIndex:  trackIndex,
-		APIURL:      h.trackAPIURL(trackID),
+		APIURL:      trackjson.TrackAPIPath(trackID),
 		FirstID:     summary.FirstID,
 		LastID:      summary.LastID,
 		MarkerCount: summary.MarkerCount,
 		Markers:     markers,
-		Disclaimers: disclaimerTexts,
+		Disclaimers: trackjson.Disclaimers,
 	}
 
 	h.respondJSON(w, resp)
 }
 
-// makeTrackMarkerPayload converts a database marker into the API representation
-// while skipping optional fields that sensors never reported.
-func makeTrackMarkerPayload(marker database.Marker) trackMarkerPayload {
-	ts, unixSeconds := normalizeMarkerTime(marker.Date)
-	payload := trackMarkerPayload{
-		ID:                     marker.ID,
-		TimeUnix:               unixSeconds,
-		TimeUTC:                ts.Format(time.RFC3339),
-		Lat:                    marker.Lat,
-		Lon:                    marker.Lon,
-		DoseRateMicroSvH:       marker.DoseRate,
-		DoseRateMicroRoentgenH: marker.DoseRate * microRoentgenPerMicroSievert,
-		CountRateCPS:           marker.CountRate,
-		SpeedMS:                marker.Speed,
-		SpeedKMH:               marker.Speed * 3.6,
-	}
-	if marker.AltitudeValid {
-		altitude := marker.Altitude
-		payload.AltitudeM = &altitude
-	}
-	if marker.TemperatureValid {
-		temperature := marker.Temperature
-		payload.TemperatureC = &temperature
-	}
-	if marker.HumidityValid {
-		humidity := marker.Humidity
-		payload.HumidityPercent = &humidity
-	}
-	detector := strings.TrimSpace(marker.Detector)
-	if detector != "" {
-		payload.DetectorType = detector
-		payload.DetectorName = stableDetectorName(marker.TrackID, detector)
-	}
-	if channels := splitRadiationChannels(marker.Radiation); len(channels) > 0 {
-		payload.RadiationTypes = channels
-	}
-	return payload
-}
-
-// handleArchiveDownload streams the daily tar.gz produced by the generator.
+// handleArchiveDownload streams the daily tar.gz bundle of .cim JSON tracks produced by the generator.
 func (h *Handler) handleArchiveDownload(w http.ResponseWriter, r *http.Request) {
 	if h.Archive == nil {
 		http.Error(w, "archive disabled", http.StatusServiceUnavailable)
@@ -726,7 +663,7 @@ func (h *Handler) finalizeSummaries(
 	for i := range summaries {
 		idx := base + int64(i) + 1
 		summaries[i].Index = idx
-		summaries[i].APIURL = h.trackAPIURL(summaries[i].TrackID)
+		summaries[i].APIURL = trackjson.TrackAPIPath(summaries[i].TrackID)
 	}
 
 	return summaries[0].Index, nil
@@ -750,107 +687,11 @@ func (h *Handler) latestTrackInfo(ctx context.Context) (int64, string, error) {
 	return total, trackID, nil
 }
 
-// trackAPIURL builds the canonical API link for a track, escaping the ID so it
-// remains safe even with unusual characters.
-func (h *Handler) trackAPIURL(trackID string) string {
-	return "/api/track/" + url.PathEscape(trackID) + ".cim"
-}
-
-// safeCIMFilename keeps attachment names predictable so browsers download
-// JSON tracks with a deterministic .cim suffix.
-func safeCIMFilename(trackID string) string {
-	var b strings.Builder
-	for _, r := range trackID {
-		switch {
-		case r >= 'a' && r <= 'z':
-			b.WriteRune(r)
-		case r >= 'A' && r <= 'Z':
-			b.WriteRune(r)
-		case r >= '0' && r <= '9':
-			b.WriteRune(r)
-		case r == '-' || r == '_':
-			b.WriteRune(r)
-		default:
-			b.WriteRune('_')
-		}
-	}
-	name := strings.Trim(b.String(), "_")
-	if name == "" {
-		name = "track"
-	}
-	return name + ".cim"
-}
-
 func (h *Handler) respondJSON(w http.ResponseWriter, payload any) {
 	w.Header().Set("Content-Type", "application/json")
 	enc := json.NewEncoder(w)
 	enc.SetIndent("", "  ")
 	_ = enc.Encode(payload)
-}
-
-func normalizeMarkerTime(unixValue int64) (time.Time, int64) {
-	if unixValue <= 0 {
-		ts := time.Unix(0, 0).UTC()
-		return ts, ts.Unix()
-	}
-	if unixValue > 1_000_000_000_000 {
-		ts := time.UnixMilli(unixValue).UTC()
-		return ts, ts.Unix()
-	}
-	ts := time.Unix(unixValue, 0).UTC()
-	return ts, unixValue
-}
-
-func splitRadiationChannels(raw string) []string {
-	trimmed := strings.TrimSpace(raw)
-	if trimmed == "" {
-		return nil
-	}
-	fields := strings.FieldsFunc(trimmed, func(r rune) bool {
-		switch r {
-		case ',', ';', '|', '/', '\\':
-			return true
-		case ' ', '\t', '\n', '\r':
-			return true
-		default:
-			return false
-		}
-	})
-	seen := make(map[string]struct{})
-	out := make([]string, 0, len(fields))
-	for _, f := range fields {
-		channel := strings.ToLower(strings.TrimSpace(f))
-		if channel == "" {
-			continue
-		}
-		if _, ok := seen[channel]; ok {
-			continue
-		}
-		seen[channel] = struct{}{}
-		out = append(out, channel)
-	}
-	if len(out) == 0 {
-		return nil
-	}
-	return out
-}
-
-// stableDetectorName synthesises a reproducible detector label. We fold the
-// track ID into the name so uploads that lacked an explicit serial number can
-// still be matched when data is re-imported later.
-func stableDetectorName(trackID, detector string) string {
-	trackID = strings.TrimSpace(trackID)
-	detector = strings.TrimSpace(detector)
-	switch {
-	case trackID == "" && detector == "":
-		return ""
-	case trackID == "":
-		return detector
-	case detector == "":
-		return trackID
-	default:
-		return trackID + ":" + detector
-	}
 }
 
 func parseIntDefault(v string, def int) int {
@@ -883,20 +724,4 @@ func clampInt(v, min, max int) int {
 		return max
 	}
 	return v
-}
-
-var disclaimerTexts = map[string]string{
-	"en": "Data is free to download. We do not create it and take no responsibility for its contents.",
-	"ru": "Данные доступны для свободного скачивания. Мы их не создаём и не несём за них никакой ответственности.",
-	"es": "Los datos se pueden descargar libremente. No los creamos y no asumimos ninguna responsabilidad por su contenido.",
-	"fr": "Les données sont libres de téléchargement. Nous ne les créons pas et n'assumons aucune responsabilité quant à leur contenu.",
-	"de": "Die Daten können frei heruntergeladen werden. Wir erstellen sie nicht und übernehmen keine Verantwortung für ihren Inhalt.",
-	"pt": "Os dados são livres para download. Não os criamos e não assumimos qualquer responsabilidade pelo seu conteúdo.",
-	"it": "I dati sono scaricabili liberamente. Non li creiamo e non ci assumiamo alcuna responsabilità per il loro contenuto.",
-	"zh": "数据可自由下载。我们不创建这些数据，对其内容不承担任何责任。",
-	"ja": "データは自由にダウンロードできます。私たちはデータを作成しておらず、その内容について一切の責任を負いません。",
-	"ar": "البيانات متاحة للتنزيل مجانًا. نحن لا ننشئها ولا نتحمل أي مسؤولية عن محتواها.",
-	"hi": "डेटा मुक्त रूप से डाउनलोड किया जा सकता है। हम इसे नहीं बनाते हैं और इसकी सामग्री के लिए कोई ज़िम्मेदारी नहीं लेते हैं।",
-	"tr": "Veriler ücretsiz olarak indirilebilir. Biz bu verileri üretmiyoruz ve içeriklerinden hiçbir sorumluluk kabul etmiyoruz.",
-	"ko": "데이터는 자유롭게 다운로드할 수 있습니다. 우리는 데이터를 만들지 않으며 그 내용에 대해 어떠한 책임도 지지 않습니다.",
 }
